requirements.txt
-i https://pypi.org/simple
aiosqlite==0.22.1; python_version >= '3.9'
alembic==1.18.1; python_version >= '3.10'
annotated-doc==0.0.4; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.12.1; python_version >= '3.9'
attrs==25.4.0; python_version >= '3.9'
authlib==1.6.6; python_version >= '3.9'
bcrypt==5.0.0; python_version >= '3.8'
certifi==2026.1.4; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.4; python_version >= '3.7'
click==8.3.1; python_version >= '3.10'
cloudpickle==3.1.2; python_version >= '3.8'
colorgram.py==1.2.0
cryptography==46.0.3; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
distro==1.9.0; python_version >= '3.6'
docstring-parser==0.17.0; python_version >= '3.8'
ecdsa==0.19.1; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
fastapi==0.123.10; python_version >= '3.8'
google-adk==1.22.1; python_version >= '3.10'
google-api-core[grpc]==2.29.0; python_version >= '3.7'
google-api-python-client==2.188.0; python_version >= '3.7'
google-auth[requests]==2.47.0; python_version >= '3.8'
google-auth-httplib2==0.3.0; python_version >= '3.7'
google-cloud-aiplatform[agent-engines]==1.134.0; python_version >= '3.9'
google-cloud-appengine-logging==1.8.0; python_version >= '3.7'
google-cloud-audit-log==0.4.0; python_version >= '3.7'
google-cloud-bigquery==3.40.0; python_version >= '3.9'
google-cloud-bigquery-storage==2.36.0; python_version >= '3.7'
google-cloud-bigtable==2.35.0; python_version >= '3.7'
google-cloud-core==2.5.0; python_version >= '3.7'
google-cloud-discoveryengine==0.13.12; python_version >= '3.7'
google-cloud-iam==2.21.0; python_version >= '3.7'
google-cloud-logging==3.13.0; python_version >= '3.7'
google-cloud-monitoring==2.29.0; python_version >= '3.7'
google-cloud-pubsub==2.34.0; python_version >= '3.7'
google-cloud-resource-manager==1.16.0; python_version >= '3.7'
google-cloud-secret-manager==2.26.0; python_version >= '3.7'
google-cloud-spanner==3.62.0; python_version >= '3.9'
google-cloud-speech==2.36.0; python_version >= '3.7'
google-cloud-storage==3.8.0; python_version >= '3.7'
google-cloud-trace==1.18.0; python_version >= '3.7'
google-crc32c==1.8.0; python_version >= '3.9'
google-genai==1.60.0; python_version >= '3.10'
google-resumable-media==2.8.0; python_version >= '3.7'
googleapis-common-protos[grpc]==1.72.0; python_version >= '3.7'
graphviz==0.21; python_version >= '3.9'
greenlet==3.3.0; python_version >= '3.10'
grpc-google-iam-v1==0.14.3; python_version >= '3.7'
grpc-interceptor==0.15.4; python_version >= '3.7' and python_version < '4.0'
grpcio==1.76.0; python_version >= '3.9'
grpcio-status==1.76.0; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httplib2==0.31.1; python_version >= '3.6'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.3; python_version >= '3.9'
idna==3.11; python_version >= '3.8'
importlib-metadata==8.7.1; python_version >= '3.9'
joblib==1.5.3; python_version >= '3.9'
jsonschema==4.26.0; python_version >= '3.10'
jsonschema-specifications==2025.9.1; python_version >= '3.9'
mako==1.3.10; python_version >= '3.8'
markdown-it-py==4.0.0; python_version >= '3.10'
markupsafe==3.0.3; python_version >= '3.9'
mcp==1.25.0; python_version >= '3.10'
mdurl==0.1.2; python_version >= '3.7'
mmh3==5.2.0; python_version >= '3.9'
numpy==2.4.1; python_version >= '3.11'
opentelemetry-api==1.37.0; python_version >= '3.9'
opentelemetry-exporter-gcp-logging==1.11.0a0; python_version >= '3.9'
opentelemetry-exporter-gcp-monitoring==1.11.0a0; python_version >= '3.9'
opentelemetry-exporter-gcp-trace==1.11.0; python_version >= '3.9'
opentelemetry-exporter-otlp-proto-common==1.37.0; python_version >= '3.9'
opentelemetry-exporter-otlp-proto-http==1.37.0; python_version >= '3.9'
opentelemetry-proto==1.37.0; python_version >= '3.9'
opentelemetry-resourcedetector-gcp==1.11.0a0; python_version >= '3.9'
opentelemetry-sdk==1.37.0; python_version >= '3.9'
opentelemetry-semantic-conventions==0.58b0; python_version >= '3.9'
packaging==26.0; python_version >= '3.8'
pandas==2.3.3; python_version >= '3.9'
passlib[bcrypt]==1.7.4
pillow==12.1.0; python_version >= '3.10'
proto-plus==1.27.0; python_version >= '3.7'
protobuf==6.33.4; python_version >= '3.9'
psycopg2-binary==2.9.11; python_version >= '3.9'
pyarrow==23.0.0; python_version >= '3.10'
pyasn1==0.6.2; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==3.0; python_version >= '3.10'
pydantic==2.12.5; python_version >= '3.9'
pydantic-core==2.41.5; python_version >= '3.9'
pydantic-settings==2.12.0; python_version >= '3.10'
pygments==2.19.2; python_version >= '3.8'
pyjwt[crypto]==2.10.1; python_version >= '3.9'
pyparsing==3.3.2; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.2.1; python_version >= '3.9'
python-jose[cryptography]==3.5.0; python_version >= '3.9'
python-multipart==0.0.21; python_version >= '3.10'
pyyaml==6.0.3; python_version >= '3.8'
referencing==0.37.0; python_version >= '3.10'
requests==2.32.5; python_version >= '3.9'
rich==14.3.1; python_full_version >= '3.8.0'
rpds-py==0.30.0; python_version >= '3.10'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
scikit-learn==1.8.0; python_version >= '3.11'
scipy==1.17.0; python_version >= '3.11'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.46; python_version >= '3.7'
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5; python_version >= '3.8'
sse-starlette==3.2.0; python_version >= '3.9'
starlette==0.50.0; python_version >= '3.10'
tenacity==9.1.2; python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
tzlocal==5.3.1; python_version >= '3.9'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.6.3; python_version >= '3.9'
uvicorn==0.40.0; python_version >= '3.10'
watchdog==6.0.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
zipp==3.23.0; python_version >= '3.9'
